Refactor common actions into separate files
Making way for additional actions
